refactoring : moving non Maven specific code outside of ArchUnitMojo, so that ArchUnitMojo is only the interface between the "Maven world" and our core logic
